Some changes to execution SOP
Moving the following out of the individual execution methods where it's repeated for each
method and merging the borgification recommendation from its own point into the body
disposal point. Slight rewording included.
"In all cases, the prisoner's pulse is to be checked to confirm death. A final round via
ballistic or energy weapon to the cranium is advised if the prisoner was not terminated."
"Executed prisoners must be borged, fired into space via mass driver, cremated, or placed
in the morgue with a DNR Notice, at the discretion of the Magistrate, Captain or Head of
Security. Borgification is recommended."
Removing checking of pulse.
Changing
"If medical personnel have been assigned to security, best effort should be made to inform
them of any performed execution so that medical records of the prisoner may be updated."
to
"Medical personnel should be informed of any performed execution, so that medical records
of the prisoner may be updated."
because brig medic isn't a thing any more.
And nesting execution methods under Execution as well as formatting them like other sub sections.